Break up Player Module
