+++ server/player.go
@@ -19,72 +19,15 @@ limitations under the License.
 package main
 
 import (
-	"bufio"
 	"context"
-	"errors"
 	"fmt"
-	"io"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
 
-	"github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
 	"github.com/gofrs/uuid/v5"
 	"golang.org/x/crypto/ssh"
 )
-
-// InputBuffer provides thread-safe operations for storing and manipulating runes
-type InputBuffer struct {
-	buffer []rune
-	mutex  sync.Mutex
-}
-
-func NewInputBuffer() *InputBuffer {
-	return &InputBuffer{
-		buffer: make([]rune, 0, 1024),
-	}
-}
-
-func (ib *InputBuffer) Append(r rune) bool {
-	ib.mutex.Lock()
-	defer ib.mutex.Unlock()
-
-	if len(ib.buffer) >= 240 {
-		return false
-	}
-	ib.buffer = append(ib.buffer, r)
-	return true
-}
-
-func (ib *InputBuffer) RemoveLast() bool {
-	ib.mutex.Lock()
-	defer ib.mutex.Unlock()
-
-	if len(ib.buffer) == 0 {
-		return false
-	}
-	ib.buffer = ib.buffer[:len(ib.buffer)-1]
-	return true
-}
-
-func (ib *InputBuffer) Clear() {
-	ib.mutex.Lock()
-	defer ib.mutex.Unlock()
-	ib.buffer = ib.buffer[:0]
-}
-
-func (ib *InputBuffer) String() string {
-	ib.mutex.Lock()
-	defer ib.mutex.Unlock()
-	return string(ib.buffer)
-}
-
-func (ib *InputBuffer) Length() int {
-	ib.mutex.Lock()
-	defer ib.mutex.Unlock()
-	return len(ib.buffer)
-}
 
 type Player struct {
 	ctx           context.Context
@@ -112,112 +55,6 @@ type Player struct {
 	shutdownOnce  sync.Once
 	done          chan struct{} // Channel signaled when all goroutines complete
 	inputBuffer   *InputBuffer  // Track current input line content
-}
-
-type PlayerData struct {
-	PlayerID      string            `json:"PlayerID" dynamodbav:"PlayerID"` // Store UUID as string in DynamoDB
-	Email         string            `json:"Email" dynamodbav:"Email"`       // Store email
-	CharacterList map[string]string `json:"CharacterList" dynamodbav:"CharacterList"`
-	SeenMotDs     []string          `json:"SeenMotD" dynamodbav:"SeenMotD"`
-}
-
-func (p *Player) Load(playerID uuid.UUID) error {
-	Logger.Debug("Loading player data", "player_id", playerID.String())
-
-	database := p.server.database
-
-	key := map[string]types.AttributeValue{
-		"PlayerID": &types.AttributeValueMemberS{Value: playerID.String()},
-	}
-
-	var playerData PlayerData
-
-	p.characterList = make(map[string]uuid.UUID)
-	p.seenMotD = make([]uuid.UUID, 0)
-
-	err := database.Get(p.server.ctx, "players", key, &playerData)
-	if err != nil {
-		if strings.Contains(err.Error(), "item not found") {
-			Logger.Info("New player", "player_id", playerID.String(), "email", p.email)
-			p.Save()
-			return nil
-		}
-		Logger.Error("Error loading player data", "error", err)
-		return err
-	}
-
-	Logger.Info("Player data loaded", "player_id", playerID.String(), "email", playerData.Email)
-
-	p.mutex.Lock()
-	defer p.mutex.Unlock()
-
-	// Update email from database
-	p.email = playerData.Email
-
-	for characterName, characterID := range playerData.CharacterList {
-		parsedUUID, err := uuid.FromString(characterID)
-		if err != nil {
-			Logger.Error("Error parsing character ID", "character_id", characterID)
-			continue
-		}
-		p.characterList[characterName] = parsedUUID
-	}
-
-	for _, motdID := range playerData.SeenMotDs {
-		motdUUID, err := uuid.FromString(motdID)
-		if err != nil {
-			Logger.Error("Error parsing MOTD ID", "motd_id", motdID)
-			continue
-		}
-		p.seenMotD = append(p.seenMotD, motdUUID)
-	}
-
-	p.lastEdited = time.Now()
-	p.lastSaved = time.Now()
-
-	return nil
-}
-
-func (p *Player) Save() error {
-	return p.SaveWithContext(p.server.ctx)
-}
-
-// SaveWithContext saves the player data with a specific context
-// This is used during shutdown to ensure saves complete even after server context is cancelled
-func (p *Player) SaveWithContext(ctx context.Context) error {
-	Logger.Info("Saving player data", "player_id", p.id.String(), "email", p.email)
-
-	database := p.server.database
-
-	playerData := PlayerData{
-		PlayerID:      p.id.String(),
-		Email:         p.email,
-		CharacterList: make(map[string]string),
-		SeenMotDs:     make([]string, len(p.seenMotD)),
-	}
-
-	// Convert character IDs to strings
-	for characterName, characterID := range p.characterList {
-		playerData.CharacterList[characterName] = characterID.String()
-	}
-
-	// Convert MOTD IDs to strings
-	for i, motdID := range p.seenMotD {
-		playerData.SeenMotDs[i] = motdID.String()
-	}
-
-	err := database.Put(ctx, "players", playerData)
-	if err != nil {
-		Logger.Error("Error saving player data", "error", err)
-		SafeSendString(p.commandOut, "Error saving player data. Please contact an administrator.\n", p.id.String())
-		return fmt.Errorf("error saving player data: %w", err)
-	}
-
-	p.mutex.Lock()
-	p.lastSaved = time.Now()
-	p.mutex.Unlock()
-
-	return nil
 }
 
 func NewPlayerSSH(server *Server, playerEmail string, conn ssh.Channel, interfaceCtx context.Context, userUUID uuid.UUID) (*Player, error) {
@@ -397,308 +234,6 @@ func (p *Player) Stop() {
 		close(p.commandIn)
 		close(p.playerError)
 	})
-}
-
-func (p *Player) handleRequests(ctx context.Context, requests <-chan *ssh.Request, done chan error) {
-	defer func() {
-		if r := recover(); r != nil {
-			Logger.Warn("Recovered in handleRequests", "player", p.id, "recover", r)
-			done <- fmt.Errorf("panic in request handler: %v", r)
-		}
-	}()
-
-	for {
-		select {
-		case <-ctx.Done():
-			done <- ctx.Err()
-			return
-		case req, ok := <-requests:
-			if !ok {
-				Logger.Debug("Request channel closed", "player", p.id)
-				done <- nil
-				return
-			}
-
-			p.mutex.Lock()
-			switch req.Type {
-			case "shell":
-				req.Reply(true, nil)
-			case "pty-req":
-				termLen := req.Payload[3]
-				w, h := ParseDims(req.Payload[termLen+4:])
-				p.consoleWidth = w
-				p.consoleHeight = h
-				req.Reply(true, nil)
-			case "window-change":
-				w, h := ParseDims(req.Payload)
-				p.consoleWidth = w
-				p.consoleHeight = h
-			default:
-				req.Reply(false, nil)
-			}
-			p.mutex.Unlock()
-		}
-	}
-}
-
-func (p *Player) handleInput(ctx context.Context, done chan error) {
-	defer func() {
-		if r := recover(); r != nil {
-			Logger.Warn("Recovered in handleInput", "player", p.id, "recover", r)
-			done <- fmt.Errorf("panic in input handler: %v", r)
-		}
-	}()
-
-	reader := bufio.NewReader(p.connection)
-
-	// Create idle timeout timer
-	idleTimeout := time.Duration(p.server.config.Game.PlayerIdleTimeoutSeconds) * time.Second
-	idleTimer := time.NewTimer(idleTimeout)
-	defer idleTimer.Stop()
-
-	// Warning timer (5 minutes before disconnect, or half the timeout if less than 10 minutes)
-	var warningTime time.Duration
-	if idleTimeout > 10*time.Minute {
-		warningTime = 5 * time.Minute
-	} else {
-		warningTime = idleTimeout / 2
-	}
-
-	var warningTimer *time.Timer
-	var warningShown bool
-
-	if warningTime > 0 {
-		warningTimer = time.NewTimer(idleTimeout - warningTime)
-		defer warningTimer.Stop()
-	}
-
-	for {
-		select {
-		case <-ctx.Done():
-			done <- ctx.Err()
-			return
-		case <-func() <-chan time.Time {
-			if warningTimer != nil {
-				return warningTimer.C
-			}
-			return nil
-		}():
-			if !warningShown {
-				warningShown = true
-				remainingMinutes := int(warningTime.Minutes())
-				select {
-				case p.commandOut <- ApplyColor("yellow", fmt.Sprintf("\r\n*** WARNING: You will be disconnected in %d minutes due to inactivity ***\r\n", remainingMinutes)):
-				case <-ctx.Done():
-					done <- ctx.Err()
-					return
-				}
-			}
-		case <-idleTimer.C:
-			select {
-			case p.commandOut <- ApplyColor("red", "\r\n*** Disconnected due to inactivity ***\r\n"):
-			case <-ctx.Done():
-			}
-			Logger.Info("Player disconnected due to idle timeout", "player", p.id)
-			done <- errors.New("idle timeout")
-			return
-		default:
-			r, _, err := reader.ReadRune()
-			if err != nil {
-				if err == io.EOF {
-					Logger.Info("Connection closed by client", "player", p.id)
-					done <- nil
-					return
-				}
-				Logger.Error("Error reading input", "player", p.id, "error", err)
-				done <- fmt.Errorf("read error: %w", err)
-				return
-			}
-
-			// Reset idle timers on any input
-			if !idleTimer.Stop() {
-				select {
-				case <-idleTimer.C:
-				default:
-				}
-			}
-			idleTimer.Reset(idleTimeout)
-
-			// Reset warning timer if not already shown
-			if !warningShown && warningTimer != nil {
-				if !warningTimer.Stop() {
-					select {
-					case <-warningTimer.C:
-					default:
-					}
-				}
-				warningTimer.Reset(idleTimeout - warningTime)
-			} else if warningShown {
-				// Clear warning state on new input
-				warningShown = false
-				if warningTimer != nil {
-					warningTimer.Reset(idleTimeout - warningTime)
-				}
-			}
-
-			// Handle different input cases
-			switch r {
-			case '\n', '\r':
-				if p.inputBuffer.Length() > 0 {
-					input := p.inputBuffer.String()
-					select {
-					case p.commandIn <- input:
-						if p.echo {
-							p.connection.Write([]byte("\r\n"))
-						}
-						p.inputBuffer.Clear()
-					case <-ctx.Done():
-						done <- ctx.Err()
-						return
-					}
-				} else {
-					// Handle empty input - just show prompt again
-					if p.echo {
-						p.connection.Write([]byte("\r\n"))
-						p.connection.Write([]byte(p.prompt))
-					}
-				}
-
-			case '\b', 127: // Backspace
-				if p.inputBuffer.RemoveLast() && p.echo {
-					p.connection.Write([]byte("\b \b"))
-				}
-
-			case '\x03': // Ctrl-C
-				done <- errors.New("player interrupt")
-				return
-
-			default:
-				// Filter input to only allow printable ASCII (32-126)
-				if r >= 32 && r <= 126 {
-					if p.inputBuffer.Append(r) && p.echo {
-						p.connection.Write([]byte(string(r)))
-					}
-				}
-			}
-		}
-	}
-}
-
-func (p *Player) handleOutput(ctx context.Context, done chan error) {
-	defer func() {
-		if r := recover(); r != nil {
-			Logger.Warn("Recovered in handleOutput", "player", p.id, "recover", r)
-			done <- fmt.Errorf("panic in output handler: %v", r)
-		}
-	}()
-
-	for {
-		select {
-		case <-ctx.Done():
-			done <- ctx.Err()
-			return
-		case msg, ok := <-p.commandOut:
-			if !ok {
-				Logger.Debug("Output channel closed", "player", p.id)
-				done <- nil
-				return
-			}
-
-			if _, err := p.connection.Write([]byte(wrapText(msg, p.consoleWidth))); err != nil {
-				Logger.Error("Write error in output handler", "player", p.id, "error", err)
-				done <- fmt.Errorf("write error: %w", err)
-				return
-			}
-		}
-	}
-}
-
-func wrapText(text string, width int) string {
-	// Handle edge cases
-	if width <= 0 {
-		width = 80 // Default width
-	}
-
-	if len(text) == 0 {
-		return text
-	}
-
-	var result strings.Builder
-	// Pre-allocate a reasonable buffer size to avoid reallocations
-	result.Grow(len(text) + len(text)/10) // Add 10% for possible line breaks
-
-	// Split the text into lines
-	lines := strings.Split(text, "\n")
-
-	for i, line := range lines {
-		// Preserve empty lines
-		if len(strings.TrimSpace(line)) == 0 {
-			result.WriteString(line)
-			if i < len(lines)-1 {
-				result.WriteString("\r\n")
-			}
-			continue
-		}
-
-		// Track position in the current line (excluding ANSI sequences)
-		linePos := 0
-		lastSpace := -1
-		startSegment := 0
-		inAnsiSequence := false
-
-		// Process each character in the line
-		for pos, char := range line {
-			// Handle ANSI escape sequences (don't count toward width)
-			if char == '\033' {
-				inAnsiSequence = true
-			}
-
-			if inAnsiSequence {
-				if char == 'm' {
-					inAnsiSequence = false
-				}
-				continue // Don't count ANSI sequence chars toward width
-			}
-
-			// Track spaces for potential line breaks
-			if char == ' ' {
-				lastSpace = pos
-			}
-
-			// Increment visible character position
-			linePos++
-
-			// Check if we need to wrap
-			if linePos > width {
-				if lastSpace != -1 {
-					// Break at the last space
-					result.WriteString(line[startSegment:lastSpace])
-					result.WriteString("\r\n")
-					startSegment = lastSpace + 1
-					linePos = pos - lastSpace
-					lastSpace = -1
-				} else {
-					// No space found, force break
-					result.WriteString(line[startSegment:pos])
-					result.WriteString("\r\n")
-					startSegment = pos
-					linePos = 1
-				}
-			}
-		}
-
-		// Add the remainder of the line
-		if startSegment < len(line) {
-			result.WriteString(line[startSegment:])
-		}
-
-		// Add line break if not the last line
-		if i < len(lines)-1 {
-			result.WriteString("\r\n")
-		}
-	}
-
-	return result.String()
 }
 
 func (s *Server) RemovePlayer(playerID uint64) error {
